Mass reworks
Fixed all possible resource leaks from db connections
Fixed servitor share buff behavior to match retail
Fixed soulshots consume
Fixed RunAway effect
Fixed some stats
Fixed premium engine duplicate exploit
Fixed some wrong calculations in pet stats
Minor codes refactor
Java quests from python
Missing messages for pet soul shots
Clean up skills from invalid item skills
Fixed hero weapon-skill stats
Fixed some wrong calculations in some skills
Fixed Blink effects not consume spiritshots
And few more i dont remember

Core changes:
compile and copy paste core.jar
Delete libs folder and paste from this rev

Data changes:
Delete xml folder and paste from this rev
Delete scripts folder and paste from this rev
Delete scripts.ini and paste from this rev

+++ L2J_SunriseProject_Core/java/l2r/gameserver/model/actor/stat/PcStat.java
@@ -663,9 +663,9 @@ public class PcStat extends PlayableStat
 	}
 	
 	@Override
-	public int getPAtkSpd()
+	public double getPAtkSpd()
 	{
-		int val = super.getPAtkSpd();
+		double val = super.getPAtkSpd();
 		
 		if ((val > FormulasConfigs.MAX_PATK_SPEED) && !getActiveChar().canOverrideCond(PcCondOverride.MAX_STATS_VALUE))
 		{
